Updated code based on comments
The score testing file is now called scoring. I modified it so you can pass a boolean to show plots (False by default). it is now exclusively a file to import from.
I made a new test folder and a new file inside it called test_scoring.py. It imports the score function and uses it the same way as before.