experimented with some ideas
Experimented with applying low/high pass filters on the sequences and visualizing the results again with a spectogram.
Also did some experimenting with cross correlation of a small section of the sequence where someone is asleep to the rest of their series. Does not look that promising but not that bad either