add event labels is refactored
Also optimized the memory usage a little more. If we split the df by series id we dont need to store the series id as long as we name the files with the series id. So I remove that column at mem_reduce and add them back at the end